requirements.txt
# FastAPI và server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# AI/LLM
openai==1.10.0
anthropic==0.18.1

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
markdown==3.5.1

# Optional: cho development
pytest==7.4.3
black==23.12.1